+++ src/run/ui_main.py
@@ -19,10 +19,15 @@ def ui_paint():
     :return:
     """
     app = QApplication(sys.argv)
+    # 设置QApplication的属性，确保当最后一个窗口关闭时应用程序也退出
+    app.setQuitOnLastWindowClosed(True)
+
+    # 连接lastWindowClosed信号到on_last_window_closed槽
+    app.lastWindowClosed.connect(on_last_window_closed)
+
     w = UIMainWindows()
     w.show()
     app.exec_()
-    app.lastWindowClosed.connect(on_last_window_closed)
 
 
 @logger.catch
@@ -31,7 +36,7 @@ def on_last_window_closed():
 
     :return:
     """
-    logger.warning("console will quit!")
+    logger.warning("Console window is closing...")
     exit_save_data()