+++ src/http_tools/JsonResponse.py
@@ -1,0 +1,32 @@
+# 导入OS模块
+import os
+import sys
+
+# 把当前文件所在文件夹的父文件夹路径加入到PYTHONPATH
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+
+class JsonResponse(object):
+    """
+    统一的json返回格式 success and error
+    """
+
+    def __init__(self, data, code, msg):
+        self.data = data
+        self.code = code
+        self.msg = msg
+
+    @classmethod
+    def success(cls, data=None, code=0, msg='success'):
+        return cls(data, code, msg)
+
+    @classmethod
+    def error(cls, data=None, code=-1, msg='error'):
+        return cls(data, code, msg)
+
+    def to_dict(self):
+        return {
+            "code": self.code,
+            "msg": self.msg,
+            "data": self.data
+        }