+++ src/log/log_record.py
@@ -26,7 +26,7 @@ def log_record():
     # 创建一个RotatingFileHandler，设置日志文件大小为1MB，备份数量为5
     logger.configure(
         handlers=[{"sink": sys.stdout, "level": sis_log_level, "format": format_record}])
-    logger.add(LOG_DIR, encoding='utf-8', rotation="00:00",  retention="30 days", compression="zip", )
+    logger.add(LOG_DIR, encoding='utf-8', rotation="00:00",  retention="30 days", compression="zip")
     logger.debug('SIS log file started record, current log file: ' + LOG_DIR)
     logging.getLogger("uvicorn.access").handlers = [InterceptHandler()]
     return True