+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add programmer quit method
 - add log zip and log save date to log files
 - add incognito explore mode
+- add fast api to all function
 
 ## v1.1.0.240320
 ### 2024/3/20